+++ src/main/java/com/example/chatbot/ObjectTypesGenerator.java
@@ -1,5 +1,7 @@
 package com.example.chatbot;
 
+import com.example.chatbot.Management.AuthHeadersManagement;
+import com.example.chatbot.Management.TokenManagement;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.http.HttpEntity;
